feat: implement `FernetEngine` encryption backend.
